Feature/incidents - Implemented

+++ Backend/src/Potholio.Core/Domain/Geolocations/Geolocation.cs
@@ -5,7 +5,7 @@ namespace Potholio.Domain.Geolocations
 {
     public class Geolocation : FullAuditedEntity<Guid>
     {
-        public virtual double Latitude { get; set; }
-        public virtual double Longitude { get; set; }
+        public virtual decimal Latitude { get; set; }
+        public virtual decimal Longitude { get; set; }
     }
 }
